Display input appended character when no value is present  (#728)
* Split input masks into separate module
