+++ components/molecules/TooltipInput/TooltipInput.js
@@ -4,7 +4,8 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import ReactTooltip from 'react-tooltip';
 import Tooltip from '~components/atoms/Tooltip/Tooltip';
-import Input, {maskEnum} from '~components/atoms/Input/Input';
+import Input from '~components/atoms/Input/Input';
+import {maskEnum} from '~components/atoms/Input/Input.masks';
 import Button from '~components/atoms/Button/Button';
 import TrashIcon from '~components/atoms/icons/TrashIcon/TrashIcon';
 import './TooltipInput.scss';
